vendor/cm: Don't add checkapi targets when invoking 'mms'
The invocation of these targets breaks the 'mms' commands that are used
for building the kernel rapidly without going through the whole
dependency tree.

This patch ignores them when running in this mode. Please note that
this is a command used for kernel development, any other commands are
unaffected by this change.

Change-Id: I8ccbef53aa540f3c25bfa83ab5989177025cb222