Revert "Update KernelAdiutor"
This is causing an issue.  Thanks to Rahul for pointing it out.
I had to compile the latest changes a certain way and for some reason
its causing the build to error out.  Revert this for now and use the
older version for now until I can figure out what is going on.

This reverts commit 4200af633fb5ea0ea08faab229844ce714850a50.